Handling Collisions Session + Intro to Pickup Programming : - 30th May 2022
PC's in school reset on 27th May so Handling Collisions with OverlapPoint brought forward to 30th.

Physics2D.OverlapPoint() functionality coded into a PointInteraction Script that sits on the PC Controller Gameobject. Currently implements a mock check point system
that detects overlap with checkpoint positions placed on the platforms in the scene.

Students assignments is to use the same functionality to create a pick up system that implements the same OverlapPoint functionality only this time the
script should sit on the pick up object and should use LayerMasks to detect overlap with the ball(PC Controller).